requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.88.0
matplotlib==3.7.1
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
requests==2.30.0
gunicorn==20.1.0
pytest==7.3.1
werkzeug==2.2.3
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.3
markupsafe==2.1.2
cachelib==0.9.0
flask-session==0.4.0
six==1.16.0
pytz==2023.3
certifi==2023.5.7
urllib3<2.0
charset-normalizer==3.1.0
idna==3.4